refactor: lift the leader-hint trailer key and decoder into tsoracle-proto (#91)
The wire-contract trailer key "tsoracle-leader-hint-bin" was hard-coded in the
server and client (plus three further test literals), and the decode helper was
duplicated across both crates. If the key ever drifted between the two sides,
the client would silently stop following NOT_LEADER hints and degrade to
round-robin.

Move the constant, the LeaderHintLookup classifier, and decode_leader_hint into
tsoracle_proto::v1 as the single source of truth. The client re-exports them so
its existing import paths are unchanged; the server keeps a thin
Option-returning adapter over the proto helper so its test-only decode surface
(and the tests covering it) stay untouched.

The richer three-way LeaderHintLookup (Absent/Decoded/Malformed) is what moves,
not the server's lossy Option<LeaderHint>, so the client keeps the
malformed-vs-absent distinction its retry loop depends on.

The literal "tsoracle-leader-hint-bin" now appears in executable code exactly
once. No behavior change on either the encode or decode path.
